Don't error if device switched off
Previously get_device or the related functions would retrieve a list of
devices that it contacted in the last x seconds. Then we would contact
each light in turn to try and apply some sort of filtering.

This would error out if the light had been switched off.

This patch attempts to contact each light and skips it if it cannot be
contacted.

This also caches the label of the light so we can display the repr
string even if the light is offline.

Closes #6